Bug 1971060 - Make virtual folders work with the new database. r=BenC The main problems here are the virtual folder view going into an infinite loop waiting for message databases to be closed, and virtual folders being removed at start-up because they don't exist in the profile like real folders. I've temporarily disabled the folder listener which updates virtual folders when the real folder hierarchy changes.
Differential Revision: https://phabricator.services.mozilla.com/D252975

--HG--
extra : rebase_source : ab819ba8cb850a78ffd7f7a0e9c3a190b04353cc
extra : histedit_source : 55ecd3f118bf642ac689b62d6ba8d17f84996564

+++ mailnews/db/panorama/src/FolderDatabase.cpp
@@ -15,6 +15,7 @@
 #include "nsIFolder.h"
 #include "nsIMsgAccountManager.h"
 #include "nsIMsgIncomingServer.h"
+#include "nsMsgFolderFlags.h"
 #include "nsThreadUtils.h"
 
 using mozilla::LazyLogModule;
@@ -408,7 +409,8 @@ NS_IMETHODIMP FolderDatabase::Reconcile(
   nsTArray<nsCString> childNames = aChildNames.Clone();
 
   for (auto child : parent->mChildren.Clone()) {
-    if (!childNames.RemoveElement(child->mName)) {
+    if (!childNames.RemoveElement(child->mName) &&
+        !(child->GetFlags() & nsMsgFolderFlags::Virtual)) {
       DeleteFolder(child);
     }
   }
